+++ pg_documentdb/include/aggregation/bson_query.h
@@ -30,4 +30,12 @@ void TraverseQueryDocumentAndProcess(bson_iter_t *queryDocument, void *context,
 									 ProcessQueryFilterFunc processFilterFunc,
 									 bool isUpsert);
 
+static inline bool
+IsBsonTypeCollationAware(bson_type_t type)
+{
+	return type == BSON_TYPE_UTF8 || type == BSON_TYPE_DOCUMENT || type ==
+		   BSON_TYPE_ARRAY;
+}
+
+
 #endif