+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### documentdb v0.103-0 (Unreleased) ###
+* Support collation with aggregation and find on sharded collections *[Feature]*
 * Support `$convert` on `binData` to `binData`, `string` to `binData` and `binData` to `string` (except with `format: auto`) *[Feature]*
 * Fix list_databases for databases with size > 2 GB *[Bugfix]* (#119)
 * Support ARM64 architecture when building docker container *[Preview]*